Move dirhash to hash package
