+++ hash/dirhash/dirhash.go
