+++ hash/dirhash/dirhash_test.go
