+++ hash/dirhash/testdata/03/symlink
