+++ hash/dirhash/testdata/04/symlink
