+++ sdk/service/service.go
@@ -11,8 +11,8 @@ import (
 	"github.com/mesg-foundation/core/container"
 	"github.com/mesg-foundation/core/database"
 	"github.com/mesg-foundation/core/hash"
+	"github.com/mesg-foundation/core/hash/dirhash"
 	"github.com/mesg-foundation/core/service"
-	"github.com/mesg-foundation/core/utils/dirhash"
 )
 
 // Service exposes service APIs of MESG.
